chore: rework config & plugin code
