+++ src/index.ts
@@ -6,3 +6,11 @@ export * from './services';
 export * from './types';
 export * from './ui';
 export * from './utils';
+
+// Services
+export { CONFIG } from './services/config/loader';
+export { type Config } from './services/config/types';
+
+export { PluginLoaderService } from './services/plugins/plugin-loader.service';
+export { type Plugin } from './services/plugins/types';
+export { definePlugin } from './services/plugins/utils';
