+++ src/services/config/types.ts
@@ -1,0 +1,6 @@
+// Types
+export interface Config {
+  jobs?: number;
+  plugins?: string[];
+  verbose?: 'info' | 'verbose' | 'debug';
+}
