+++ src/services/index.ts
@@ -1,4 +1,3 @@
-export * from './config.service';
 export * from './inversify.config';
 export * from './logger.service';
 export * from './task-expr.service';
